Use Jsonnet to define mixin dashboards
Created a GitHub repository dashboard, which graphs exporter metrics for
a given repository. Also created an API usage dashboard to check the
if the exporter is running out of requests.

The dashboards use Grafonnet, a library for writing Grafana dashboards.
Added jsonnet-bundler integration so the Grafonnet package can be
fetched.